add profile column in users & operators

+++ app/Http/Services/CategoryService.php
@@ -9,26 +9,29 @@ use Exception;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Log;
 
-class CategoryService extends BaseController{
+class CategoryService extends BaseController
+{
     public function __construct(
         private Category $category,
         private Product $product,
-        )
-    {
+    ) {
     }
 
     /////////////////////////////////////////////////////////////////////////////////////
 
     public function index($request)
     {
-        $data = Category::with(["branch" => function($q){
-            $q->select('id','name',
-        );
-      },
-        "parent" => function($query){
-            $query->select('id','name');
-        }
-        ])->orderBy('name','asc')->paginate($request['row_count']);
+        $data = Category::with([
+            "branch" => function ($q) {
+                $q->select(
+                    'id',
+                    'name',
+                );
+            },
+            "parent" => function ($query) {
+                $query->select('id', 'name');
+            }
+        ])->orderBy('name', 'asc')->paginate($request['row_count']);
         return $this->sendResponse('Caregory Index Success', $data);
     }
 
@@ -36,14 +39,15 @@ class CategoryService extends BaseController{
 
     public function store(array $request)
     {
-        if($request['parent_id'] == 'null'){
-            $request['parent_id'] =null;
+        info($request);
+        if ($request['parent_id'] == 'null') {
+            $request['parent_id'] = null;
         }
-        if(isset($request['image'])){
+        if (isset($request['image'])) {
             $image = $request['image'];
             $destinationpath = 'images/categories';
             $categoryImage = date('YmdHis') . '.' . $image->getClientOriginalExtension();
-            $image->move($destinationpath,$categoryImage);
+            $image->move($destinationpath, $categoryImage);
             $request['image'] = "$categoryImage";
         }
         $this->insertData($request, 'categories');
@@ -55,12 +59,13 @@ class CategoryService extends BaseController{
     public function edit($request)
     {
         $data = $this->category->where('id', $request['id'])
-            ->with(["branch" => function($q){
-                $q->select('id','name');
-            },
-            "parent" => function ($query) {
-                $query->select('id', 'name');
-            }
+            ->with([
+                "branch" => function ($q) {
+                    $q->select('id', 'name');
+                },
+                "parent" => function ($query) {
+                    $query->select('id', 'name');
+                }
             ])
             ->whereNull('deleted_at')->first();
         if (!$data) {
@@ -73,26 +78,29 @@ class CategoryService extends BaseController{
 
     public function update(array $request)
     {
-        // dd($request);
         try {
             $this->beginTransaction();
-        $prevImage = $this->category->where('id', $request['id'])->whereNull('deleted_at')->value('image');
-        if (isset($request['image'])) {
-            if ($request['image'] != '') {
-                $path = 'images/categories/' . $prevImage;
-                if (File::exists($path)) {
-                    File::delete($path);
+            $prevImage = $this->category->where('id', $request['id'])->whereNull('deleted_at')->value('image');
+            if (isset($request['image'])) {
+                if (str_contains($request['image'], 'C:\xampp\tmp')) {
+                    info(['image type string']);
+                    if ($request['image'] != '') {
+                        $path = 'images/categories/' . $prevImage;
+                        if (File::exists($path)) {
+                            File::delete($path);
+                        }
+                    }
+                    $image = $request['image'];
+                    $destinationpath = 'images/categories';
+                    $categoryImage = date('YmdHis') . '.' . $image->getClientOriginalExtension();
+                    $image->move($destinationpath, $categoryImage);
+                    $request['image'] = "$categoryImage";
                 }
+
             }
-                $image = $request['image'];
-                $destinationpath = 'images/categories';
-                $categoryImage = date('YmdHis') . '.' . $image->getClientOriginalExtension();
-                $image->move($destinationpath, $categoryImage);
-                $request['image'] = "$categoryImage";
-        }
-        $this->updateData($request, 'categories');
-        $this->commit();
-        return $this->sendResponse('Category Update Success');
+            $this->updateData($request, 'categories');
+            $this->commit();
+            return $this->sendResponse('Category Update Success');
         } catch (\Exception $e) {
             $this->rollback();
             throw new \Exception($e);
@@ -103,7 +111,7 @@ class CategoryService extends BaseController{
 
     public function delete($request)
     {
-        try{
+        try {
             $this->beginTransaction();
             $id = $this->category->find($request['id']);
             if (!$id) {
@@ -114,8 +122,8 @@ class CategoryService extends BaseController{
             if (File::exists($path)) {
                 File::delete($path);
             }
-            $product  = $this->product->where('category_id',$request['id'])->first();
-            $category  = $this->category->where('parent_id',$request['id'])->first();
+            $product  = $this->product->where('category_id', $request['id'])->first();
+            $category  = $this->category->where('parent_id', $request['id'])->first();
 
             if ($product || $category) {
                 return $this->sendError("This Category is already used. Can't delete!!");
@@ -123,11 +131,10 @@ class CategoryService extends BaseController{
             $this->deleteById($request['id'], 'categories');
             $this->commit();
             return $this->sendResponse('Category Delete Success');
-        }catch(\Exception $e){
+        } catch (\Exception $e) {
             $this->rollback();
             throw new \Exception($e);
         }
-
     }
 
     /////////////////////////////////////////////////////////////////////////////////////
@@ -138,14 +145,15 @@ class CategoryService extends BaseController{
         $rowCount = $request['row_count'];
         $rowCount = !$rowCount ? null : $rowCount;
         $data = Category::where('name', 'like', "%$keyword%")
-            ->with(["branch" => function($q){
-                $q->select('id','name');
-            },
-            "parent" => function ($query) {
-                $query->select('id', 'name');
-            }
+            ->with([
+                "branch" => function ($q) {
+                    $q->select('id', 'name');
+                },
+                "parent" => function ($query) {
+                    $query->select('id', 'name');
+                }
             ])
-            ->orderBy('name','asc')
+            ->orderBy('name', 'asc')
             ->paginate($rowCount);
         return $this->sendResponse('Category Search Success', $data);
     }
